requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0
onnxruntime>=1.17.0
mediapipe>=0.10.13
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4 